层级修改

+++ magic-user-service/src/main/java/com/magic/user/dao/MemberMongoDaoImpl.java
@@ -165,7 +165,7 @@ public class MemberMongoDaoImpl extends BaseMongoDAOImpl<MemberConditionVo> {
         return query;
     }
 
-    public boolean updateLevel(long memberId, int level) {
+    public boolean updateLevel(long memberId, long level) {
         Query query = new Query(new Criteria("memberId").is(memberId));
         Update update = new Update().set("level", level);
         return super.update(query, update) != null;
